+++ bindings/src/account_history/msg/mod.rs
@@ -51,8 +51,6 @@ pub mod query_resp {
 
     pub mod masterchef {
         mod get_masterchef_pending_rewards;
-        mod get_masterchef_claim_rewards;
         pub use get_masterchef_pending_rewards::GetMasterchefUserPendingRewardResponse;
-        pub use get_masterchef_claim_rewards::GetMasterchefClaimRewardsResponse;
     }
 }
