--- bindings/src/account_history/msg/query_resp/masterchef/get_masterchef_claim_rewards.rs
@@ -1,8 +1,0 @@
-use cosmwasm_schema::cw_serde;
-
-#[cw_serde]
-#[derive(Default)]
-pub struct GetMasterchefClaimRewardsResponse {
-    pub code: u64,
-    pub result: String,
-}
