+++ bindings/src/trade_shield/msg/execute_msg.rs
@@ -115,4 +115,7 @@ pub enum ExecuteMsg {
         leverage_enabled: Option<bool>,
         limit_process_order: Option<u128>, // set to zero set the limit to None
     },
+    MasterchefClaimRewards {
+        pool_ids: Vec<u64>,
+    },
 }
