+++ contracts/trade-shield-contract/src/action/mod.rs
@@ -59,6 +59,7 @@ pub mod execute {
 
     mod close_leveragelp_position_request;
     mod open_leveragelp_position_request;
+    mod masterchef_claim_rewards;
 
     use super::*;
 
@@ -83,6 +84,7 @@ pub mod execute {
     pub use open_leveragelp_position_request::open_leveragelp_position_request;
     pub use stake_request::stake_request;
     pub use unstake_request::unstake_request;
+    pub use masterchef_claim_rewards::masterchef_claim_rewards;
 }
 
 pub mod reply {
