+++ contracts/trade-shield-contract/src/entry_point/execute.rs
@@ -207,5 +207,7 @@ pub fn execute(
             }
             Ok(Response::new())
         }
+
+        MasterchefClaimRewards { pool_ids } => masterchef_claim_rewards(info, pool_ids),
     }
 }
